requirements.txt
asgiref==3.7.2
Django==4.2.1
django-credit-cards==0.4.1
django-debug-toolbar==4.1.0
django-environ==0.10.0
django-filter==23.2
django-phonenumber-field==7.1.0
django-recaptcha==3.0.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
phonenumberslite==8.13.14
Pillow==9.5.0
psycopg2-binary==2.9.6
pytz==2023.3
signals==0.0.2
sqlparse==0.4.4
typing_extensions==4.6.2